-broke client.py functions into a seperate file -had client_functions support encrypting files -transfered a test in client.py into tests.py
